Add the capability to analyze wave height and 10-m wind gust into HRRR-3DRTMA package. 1. GSI source code would be git-cloned from github repository; 2. Add steps in workflow (ex-scripts, j-job scripts, xml files; 3. some updated fix files (anavonfo,convinfo under fix/gsi) and parm file    (gsiparm.anl.sh under parm/gsi) are NOT included in this pacakge.

+++ sorc/build_rtma_gsi.sh
@@ -25,16 +25,33 @@ setenv LDFLAGS_COM " "
 
 module list
 
-cd ${BASE}/rtma_gsi.fd
-rm -fr build
-mkdir build
-cd ush
-./build_mine.sh
-#cmake -DENKF_MODE=WRF -DBUILD_ENKF_PREPROCESS_ARW=ON -DBUILD_GSDCLOUD_ARW=ON ../.
-#make -j1
+cd ${BASE}
+[[ -d ./rtma_gsi.fd ]] && rm -rf ./rtma_gsi.fd
 
-cp ${BASE}/rtma_gsi.fd/build/src/gsi/gsi.x        ${BASE}/../exec/rtma_gsi
-#cpreq bin/enkf_wrf.x   ${BASE}/../exec/rtma_enkf
+# get the gsi source code package from repository on github
+  echo "git clone https://github.com/NOAA-EMC/GSI.git  ./rtma_gsi.fd"
+  git clone https://github.com/NOAA-EMC/GSI.git  ./rtma_gsi.fd
+  cd ${BASE}/rtma_gsi.fd
+  echo "git checkout develop"
+  git checkout develop       # <--- checking out the latest commit of develop branch
+# git checkout 0ef8d87       # <--- specifying the commit with updates for analysis fof howv and gust
+
+  [[ -d ./build ]] && rm -fr build
+  mkdir build
+
+  cd ush
+#  use the gsi building script
+  if [[ -f ${BASE}/../util_dev/build_gsi4rtma3d.sh ]] ; then
+     cp -p ${BASE}/../util_dev/build_gsi4rtma3d.sh  ./
+    ./build_gsi4rtma3d.sh
+  else
+     echo "cannot find the GSI building script build_gsi4rtma3d.sh, abort build GSI ..."
+  fi
+
+cp -p ${BASE}/rtma_gsi.fd/build/src/gsi/gsi.x        ${BASE}/../exec/rtma_gsi
+# cp -p ${BASE}/rtma_gsi.fd/install/bin/gsi.x        ${BASE}/../exec/rtma_gsi
+cp -p ${BASE}/rtma_gsi.fd/build/src/enkf/enkf.x      ${BASE}/../exec/rtma_enkf
+# cp -p ${BASE}/rtma_gsi.fd/install/bin/enkf.x       ${BASE}/../exec/rtma_enkf
 #cpreq bin/enspreproc.x ${BASE}/../exec/rtma_process_enkf
 #cpreq bin/initialens.x ${BASE}/../exec/rtma_initialens
